HBX-1232: Add gradle support   - Pull up following methods from 'GenerateCfgTask' and 'GenerateJavaTask' into AbstractTask     * createJdbcDescriptor()     * getOutputFolder()     * setupReverseEngineeringStrategy()   - Move test 'GenerateJavaTaskTest#testSetupReverseEngineeringStrategy()' to 'AbstractTaskTest'   - Remove unneeded test class 'GenerateJavaTaskTest'
